register bundles with server
so server can cleanup memory when a client closes a bundle or a bundle becomes "stale" without interaction for a set amount of time.  
Popout windows now automatically close themselves when detaching from the server (either by closing the window, changing bundles, changing servers).  In the future, we may want to loosen this for a server change... but then we'll need to communicate the new server (and possibly bundleid) to each window and have them automatically update from that information.